feat(ui): refactor Year component and update page.tsx
