+++ apps/blog/app/[year]/page.tsx
@@ -1,7 +1,6 @@
 import YearList from '@duyet/components/YearList';
-import type { Post } from '@duyet/interfaces';
-import { getPostsByAllYear, getPostsByYear } from '@duyet/libs/getPost';
-import Link from 'next/link';
+import { getPostsByAllYear } from '@duyet/libs/getPost';
+import Year from '../../components/Year';
 
 // Dynamic segments not included in generateStaticParams will return a 404.
 // https://nextjs.org/docs/app/api-reference/file-conventions/route-segment-config#dynamicparams
@@ -13,25 +12,10 @@ interface YearProps {
   };
 }
 
-export default function Year({ params: { year } }: YearProps) {
-  const posts = getPostsByYear(year, ['slug', 'title', 'date', 'category']);
-
+export default function YearPage({ params: { year } }: YearProps) {
   return (
     <>
-      <h1 className="mb-5 mt-10 text-3xl font-bold">{year}</h1>
-
-      {posts.map((post: Post) => (
-        <article className="mb-5" key={post.slug}>
-          <div className="mb-2 flex flex-row gap-2">
-            <time className="text-gray-400">{post.date.toString()}</time>
-            <span className="text-gray-500">{post.category}</span>
-          </div>
-
-          <Link as={post.slug} className="text-xl font-bold" href="/[...slug]">
-            {post.title}
-          </Link>
-        </article>
-      ))}
+      <Year year={year} />
 
       <div className="mt-10">
         <YearList />
