+++ apps/blog/app/archives/page.tsx
@@ -1,9 +1,7 @@
 import Container from '@duyet/components/Container';
 import YearList from '@duyet/components/YearList';
-import type { Post } from '@duyet/interfaces';
 import { getPostsByAllYear } from '@duyet/libs/getPost';
-import { cn } from '@duyet/libs/utils';
-import Link from 'next/link';
+import Year from '../../components/Year';
 
 export default function Archives() {
   const yearLimit = 5;
@@ -17,44 +15,9 @@ export default function Archives() {
       <div>
         {Object.keys(postsByYear)
           .sort((a: string, b: string) => parseInt(b) - parseInt(a))
-          .map((year: string) => {
-            const posts = postsByYear[parseInt(year)];
-
-            return (
-              <div key={year}>
-                <Link as={`/${year}`} href="/[year]">
-                  <h1
-                    className={cn(
-                      'mb-8 mt-8 text-5xl font-extrabold',
-                      'sm:text-6xl',
-                      'md:mb-10 md:mb-10 md:text-8xl md:font-black',
-                    )}
-                  >
-                    {year}
-                  </h1>
-                </Link>
-
-                {posts.map((post: Post) => (
-                  <article className="mb-5" key={post.slug}>
-                    <div className="mb-2 flex flex-row gap-2">
-                      <time className="text-gray-400">
-                        {post.date.toString()}
-                      </time>
-                      <span className="text-gray-500">{post.category}</span>
-                    </div>
-
-                    <Link
-                      as={post.slug}
-                      className="text-xl font-bold md:text-2xl"
-                      href="/[...slug]"
-                    >
-                      {post.title}
-                    </Link>
-                  </article>
-                ))}
-              </div>
-            );
-          })}
+          .map((year: string) => (
+            <Year key={year} year={parseInt(year)} />
+          ))}
       </div>
 
       <div className="border-top-1 mt-10">
